split out the Configurable implementation so it can save the settings correctly, as the Configurable is created fresh each settings dialog open
